+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/Block.kt
@@ -1,0 +1,3 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class Block(val confidence: Double, val text: String)
