+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/DocumentMetadata.kt
@@ -1,0 +1,3 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class DocumentMetadata(val pages: Int)
