+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/LineItem.kt
@@ -1,0 +1,3 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class LineItem(val lineItemExpenseFields: Array<LineItemExpenseField>)
