+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/LineItemExpenseField.kt
@@ -1,0 +1,8 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class LineItemExpenseField(
+    val productCode: SummaryField?,
+    val item: SummaryField?,
+    val price: SummaryField?,
+    val expenseRow: SummaryField?
+)
