+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/LineItemGroup.kt
@@ -1,0 +1,3 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class LineItemGroup(val lineItemGroupIndex: Int, val lineItems: Array<LineItem>)
