+++ client/src/main/java/com/mytiki/publish/client/capture/rsp/SummaryField.kt
@@ -1,0 +1,3 @@
+package com.mytiki.publish.client.capture.rsp
+
+data class SummaryField(val confidenceKey: Int, val confidenceValue: Int, val value: String)
